In to 3D

+++ YasEngine/time_picker.cpp
@@ -9,7 +9,7 @@ TimePicker::TimePicker() { ; }
 
 double TimePicker::GetSeconds() { return SDL_GetTicks() / 1000.0; }
 
-double TimePicker::GetMiliseconds() { return SDL_GetTicks(); }
+double TimePicker::GetMiliseconds() { return static_cast<double>(SDL_GetTicks()); }
 //                                                                            80
 //                                                                            120
 //-----------------------------------------------------------------------------|---------------------------------------|
